dwgTs-parity: decode VISUALSTYLE object body (DWG)
Replace the no-op DRW_VisualStyle::parseDwg stub with the full
AcDbVisualStyle body decoder covering all three layouts:

- legacy (< AC1024)   -- readVisualStyleLegacyBody (dwgTs 1587-1627)
- r2010b (>= AC1024)  -- readVisualStyleR2010bBody (dwgTs 1762-1832)
- r2013b (>= AC1027)  -- readVisualStyleR2013bExpansion (dwgTs 1651-1710)

Adds a DRW_VisualStyleBody struct holding face/edge/display groups +
r2010b/r2013b extensions, mirrored 1:1 from dwgTs parseObjects.ts and
cross-checked against libreDWG dwg2.spec:2119-2289. Widens
DRW_VisualStyle::type BS -> BL (dwg2.spec: style_type is BL 70) and
mirrors the same widening in LC_DwgAdvancedMetadata::VisualStyleRecord
to keep the record.type = visualStyle.type assignment non-narrowing.

Implementation notes:
- parseDwg always returns true once the common preamble parses, so
  the dispatcher's `if(ret)` in dwgreader.cpp:2541 still delivers
  BOTH addVisualStyle and the raw shelf (graceful degrade).
- Tail guard uses numRemainingBytes() (matches dwgTs buf.remaining())
  for pre-R2007; R2007+ uses currentObjectDwgBit(buf) < objSize.
- m_body has no filter consumer today (rs_filterdxfrw.cpp:1492 only
  copies handle/parent/name/desc/type into VisualStyleRecord); this
  phase is DRW-layer structured-metadata parity with dwgTs, hence the
  [parity] tag. Round-trip already covered by the raw shelf.

Tests (new file visualstyle_dwg_tests.cpp + CMake registration):
- r2013b/r2010b: Point.dwg (AC1027) -- 30+ asserts including the
  face_opacity=+0.6 branch discriminator, r2013b prop* fields, and
  strokes="strokes_ogs.tif". Fully asserts.
- legacy: Point_1.dwg (AC1021) -- graceful SUCCEED-skip when the
  pre-existing R2007 BAD_READ_HEADER hazard prevents the read (same
  class as the 7 accepted dwg_smoke_tests failures on master). Fires
  automatically once that upstream hazard is fixed; the P3.3 legacy
  decode itself is byte-exact per dwgTs.

Fixtures visualstyle_r2013.dwg / visualstyle_r2007.dwg copied into
testdata from ~/doc/dwg6/Point.dwg / Point_1.dwg.

Plan: librecad-match-dwgts-round2, PR P3.3 (VISUALSTYLE body decode)

+++ librecad/src/lib/engine/document/lc_dwgadvancedmetadata.h
@@ -743,7 +743,7 @@ public:
         std::uint32_t parentHandle = 0;
         std::string name;
         std::string description;
-        std::uint16_t type = 0;
+        std::int32_t type = 0;   // mirrors DRW_VisualStyle::type widening BS->BL (P3.3)
         ReplayState replayState = ReplayState::ReplayAllowed;
     };
